+++ src/internals.ts
@@ -17,6 +17,56 @@ export { Shape } from './Shape';
 export { Animation } from './Animation';
 export { Tween, Easings } from './Tween';
 
+export const enableTrace = false;
+export const traceArrMax = 100;
+export const listenClickTap = false;
+export const inDblClickWindow = false;
+
+/**
+ * Global pixel ratio configuration. KonvaJS automatically detect pixel ratio of current device.
+ * But you may override such property, if you want to use your value.
+ * @property pixelRatio
+ * @default undefined
+ * @name pixelRatio
+ * @memberof Konva
+ * @example
+ * Konva.pixelRatio = 1;
+ */
+export const pixelRatio = undefined;
+
+/**
+ * Drag distance property. If you start to drag a node you may want to wait until pointer is moved to some distance from start point,
+ * only then start dragging. Default is 3px.
+ * @property dragDistance
+ * @default 0
+ * @memberof Konva
+ * @example
+ * Konva.dragDistance = 10;
+ */
+export const dragDistance = 3;
+/**
+ * Use degree values for angle properties. You may set this property to false if you want to use radiant values.
+ * @property angleDeg
+ * @default true
+ * @memberof Konva
+ * @example
+ * node.rotation(45); // 45 degrees
+ * Konva.angleDeg = false;
+ * node.rotation(Math.PI / 2); // PI/2 radian
+ */
+export const angleDeg = true;
+/**
+ * Show different warnings about errors or wrong API usage
+ * @property showWarnings
+ * @default true
+ * @memberof Konva
+ * @example
+ * Konva.showWarnings = false;
+ */
+export const showWarnings = true;
+
+// export default KonvaInternals;
+
 // shapes
 export { Arc } from './shapes/Arc';
 export { Arrow } from './shapes/Arrow';